REF: extract indenting module
